+++ frontend/tweet/src/components/Trends/Trend.js
