+++ frontend/tweet/src/components/Trends/TrendSearch.js
