+++ frontend/tweet/src/components/Trends/TrendTopic.js
